Move delegates into src/delegates, delete elf-council-delegates directory
