+++ apps/council-frontend/src/delegates/isFeaturedDelegate.ts
@@ -1,4 +1,4 @@
-import { Delegate, delegates } from "src/elf-council-delegates/delegates";
+import { Delegate, delegates } from "src/delegates/delegates";
 
 export function getFeaturedDelegate(
   delegateAddress: string,
