--- apps/council-frontend/src/elf-council-delegates/README.md
@@ -1,1 +1,0 @@
-TODO: Move this to it's own project in packages/council-delegates
