+++ apps/council-frontend/src/ui/delegate/DelegatesList/DetailedDelegateProfile.tsx
@@ -1,7 +1,7 @@
 import { ReactElement, useEffect, useRef } from "react";
 import { t } from "ttag";
 import H2 from "src/ui/base/H2/H2";
-import { Delegate } from "src/elf-council-delegates/delegates";
+import { Delegate } from "src/delegates/delegates";
 import { WalletJazzicon } from "src/ui/wallet/WalletJazzicon";
 import classNames from "classnames";
 import Button from "src/ui/base/Button/Button";
